Updated code for checking valid BST and readme
